Update the docker-compose.yml (#705)
The existing docker file was a little confusing, I didn't realize you had to set the variables in an external .env file, I think it makes more sense to put it in the docker file directly, and import the env values, not the entire structure.  Also added a few quality of life improvements.  Since I am on v3.7 for docker-compose I used that version, but I believe it will work with older versions.