add a process internal filewatcher to re-start the server in dev mode
Added a --dev flag to saltcorn server to watch for file changes programmatically.
A file change will compile the typescipt files and kill the server. The watcher assumes
that pm2 will do the actual re-start and the node_modules directory of the project root
must be installed.

The pm2 command could look like:

pm2 start "`which saltcorn` serve --dev"