Set browser locale as active if !user.language
Set browser locale as active if user.language is not set yet. Otherwise, the default selection will always be English. As the form is saved on changes only, a user with a browser locale different than English has to select a different language first, before he can select and save English as language. This is quite inconvenient.